コメント指摘内容の修正完了(1)

+++ app/controllers/items_controller.rb
@@ -35,8 +35,10 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
-    item.destroy
-    redirect_to root_path
+    if user_signed_in? && current_user.id == item.user_id
+      item.destroy
+      redirect_to root_path
+    end
   end
 
   private
